# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "UniAssist API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/tasks/{task_id}/submissions": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Task Submissions",
        "operationId": "task_submissions_tasks__task_id__submissions_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          },
          {
            "name": "latest_only",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Latest Only"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/SubmissionStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionQueueItem"
                  },
                  "title": "Response Task Submissions Tasks  Task Id  Submissions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/submissions/": {
      "post": {
        "tags": [
          "Submissions"
        ],
        "summary": "Submit",
        "description": "Student submits work for a task.",
        "operationId": "submit_submissions__post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionCreatedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/submissions/my/{task_id}": {
      "get": {
        "tags": [
          "Submissions"
        ],
        "summary": "My Submissions",
        "description": "Student views all their submissions for a task.",
        "operationId": "my_submissions_submissions_my__task_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SubmissionListItemResponse"
                  },
                  "title": "Response My Submissions Submissions My  Task Id  Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/submissions/{submission_id}/grade": {
      "post": {
        "tags": [
          "Submissions"
        ],
        "summary": "Grade",
        "description": "Staff triggers AI grading for a submission.",
        "operationId": "grade_submissions__submission_id__grade_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionGradedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/submissions/{submission_id}/confirm": {
      "patch": {
        "tags": [
          "Submissions"
        ],
        "summary": "Confirm",
        "description": "Professor confirms or overrides the AI-suggested grade.",
        "operationId": "confirm_submissions__submission_id__confirm_patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Submission Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmGradeRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionConfirmedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/submissions/{submission_id}": {
      "get": {
        "tags": [
          "Submissions"
        ],
        "summary": "Get One",
        "description": "Get submission details including grade and code reviews.",
        "operationId": "get_one_submissions__submission_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "submission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Submission Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmissionDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ArtifactResponse": {
        "properties": {
          "file_id": {
            "type": "integer",
            "title": "File Id"
          },
          "file_name": {
            "type": "string",
            "title": "File Name"
          },
          "file_type": {
            "type": "string",
            "title": "File Type"
          }
        },
        "type": "object",
        "required": [
          "file_id",
          "file_name",
          "file_type"
        ],
        "title": "ArtifactResponse"
      },
...
      "CodeFindingResponse": {
        "properties": {
          "file_path": {
            "type": "string",
            "title": "File Path"
          },
          "line_number": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line Number"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "finding": {
            "type": "string",
            "title": "Finding"
          }
        },
        "type": "object",
        "required": [
          "file_path",
          "severity",
          "finding"
        ],
        "title": "CodeFindingResponse"
      },
...
      "ConfirmGradeRequest": {
        "properties": {
          "final_grade": {
            "type": "number",
            "minimum": 0,
            "title": "Final Grade"
          }
        },
        "type": "object",
        "required": [
          "final_grade"
        ],
        "title": "ConfirmGradeRequest"
      },
...
      "CreateSubmissionRequest": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "submission_text": {
            "type": "string",
            "title": "Submission Text",
            "default": ""
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          },
          "repository_snapshot_id": {
            "anyOf": [
              {
                "type": "integer",
                "exclusiveMinimum": 0
              },
              {
                "type": "null"
              }
            ],
            "title": "Repository Snapshot Id"
          }
        },
        "type": "object",
        "required": [
          "task_id"
        ],
        "title": "CreateSubmissionRequest"
      },
...
      "CriterionEvaluationResponse": {
        "properties": {
          "criterion_name": {
            "type": "string",
            "title": "Criterion Name"
          },
          "score_given": {
            "type": "number",
            "title": "Score Given"
          },
          "max_points": {
            "type": "number",
            "title": "Max Points"
          },
          "reasoning": {
            "type": "string",
            "title": "Reasoning"
          }
        },
        "type": "object",
        "required": [
          "criterion_name",
          "score_given",
          "max_points",
          "reasoning"
        ],
        "title": "CriterionEvaluationResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RosterMember": {
        "properties": {
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "student_number": {
            "type": "string",
            "title": "Student Number"
          },
          "accepted_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accepted At"
          }
        },
        "type": "object",
        "required": [
          "student_id",
          "name",
          "student_number",
          "accepted_at"
        ],
        "title": "RosterMember"
      },
...
      "RubricCriterionInput": {
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "max_points": {
            "type": "number",
            "exclusiveMinimum": 0,
            "title": "Max Points"
          },
          "sort_order": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sort Order"
          }
        },
        "type": "object",
        "required": [
          "name",
          "max_points"
        ],
        "title": "RubricCriterionInput"
      },
...
      "SnapshotInfo": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "repository_id": {
            "type": "integer",
            "title": "Repository Id"
          },
          "commit_sha": {
            "type": "string",
            "title": "Commit Sha"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "provenance": {
            "$ref": "#/components/schemas/SnapshotProvenance"
          }
        },
        "type": "object",
        "required": [
          "id",
          "repository_id",
          "commit_sha",
          "created_at",
          "provenance"
        ],
        "title": "SnapshotInfo"
      },
...
      "SubmissionConfirmedResponse": {
        "properties": {
          "submission_id": {
            "type": "integer",
            "title": "Submission Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "final_grade": {
            "type": "number",
            "title": "Final Grade"
          },
          "confirmed_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed By"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "status",
          "final_grade"
        ],
        "title": "SubmissionConfirmedResponse"
      },
      "SubmissionCreatedResponse": {
        "properties": {
          "submission_id": {
            "type": "integer",
            "title": "Submission Id"
          },
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "task_id",
          "attempt_number",
          "status",
          "submitted_at"
        ],
        "title": "SubmissionCreatedResponse"
      },
      "SubmissionDetailResponse": {
        "properties": {
          "repository_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SnapshotInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "team_snapshot": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TeamSnapshot"
              },
              {
                "type": "null"
              }
            ]
          },
          "grading_guidance_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grading Guidance Id"
          },
          "grading_guidance_version": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Grading Guidance Version"
          },
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "team_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Team Id"
          },
          "file_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Id"
          },
          "submission_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Submission Text"
          },
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number"
          },
          "is_latest": {
            "type": "boolean",
            "title": "Is Latest"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ai_suggested_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Suggested Grade"
          },
          "final_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Grade"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "confirmed_by": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed By"
          },
          "confirmed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Confirmed At"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "rubric_id": {
            "type": "integer",
            "title": "Rubric Id"
          },
          "rubric_version": {
            "type": "integer",
            "title": "Rubric Version"
          },
          "total_possible_grade": {
            "type": "number",
            "title": "Total Possible Grade"
          },
          "rubric_criteria": {
            "items": {
              "$ref": "#/components/schemas/RubricCriterionInput"
            },
            "type": "array",
            "title": "Rubric Criteria"
          },
          "artifacts": {
            "items": {
              "$ref": "#/components/schemas/ArtifactResponse"
            },
            "type": "array",
            "title": "Artifacts"
          },
          "criterion_evaluations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CriterionEvaluationResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Criterion Evaluations"
          },
          "code_reviews": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/CodeFindingResponse"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code Reviews"
          },
          "ai_warnings": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Warnings"
          },
          "is_mock": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Is Mock"
          }
        },
        "type": "object",
        "required": [
          "id",
          "task_id",
          "student_id",
          "attempt_number",
          "is_latest",
          "status",
          "submitted_at",
          "rubric_id",
          "rubric_version",
          "total_possible_grade"
        ],
        "title": "SubmissionDetailResponse"
      },
...
      "SubmissionGradedResponse": {
        "properties": {
          "submission_id": {
            "type": "integer",
            "title": "Submission Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "ai_suggested_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Suggested Grade"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          }
        },
        "type": "object",
        "required": [
          "submission_id",
          "status"
        ],
        "title": "SubmissionGradedResponse"
      },
      "SubmissionListItemResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number"
          },
          "is_latest": {
            "type": "boolean",
            "title": "Is Latest"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "ai_suggested_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Suggested Grade"
          },
          "final_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Grade"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "total_possible_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Possible Grade"
          }
        },
        "type": "object",
        "required": [
          "id",
          "attempt_number",
          "is_latest",
          "status",
          "submitted_at"
        ],
        "title": "SubmissionListItemResponse"
      },
      "SubmissionQueueItem": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "attempt_number": {
            "type": "integer",
            "title": "Attempt Number"
          },
          "is_latest": {
            "type": "boolean",
            "title": "Is Latest"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time",
            "title": "Submitted At"
          },
          "ai_suggested_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ai Suggested Grade"
          },
          "final_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Final Grade"
          },
          "feedback": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback"
          },
          "total_possible_grade": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Possible Grade"
          },
          "student_id": {
            "type": "integer",
            "title": "Student Id"
          },
          "student_name": {
            "type": "string",
            "title": "Student Name"
          },
          "student_number": {
            "type": "string",
            "title": "Student Number"
          }
        },
        "type": "object",
        "required": [
          "id",
          "attempt_number",
          "is_latest",
          "status",
          "submitted_at",
          "student_id",
          "student_name",
          "student_number"
        ],
        "title": "SubmissionQueueItem"
      },
      "SubmissionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "ai_graded",
          "staff_confirmed"
        ],
        "title": "SubmissionStatus"
      },
...
      "TeamSnapshot": {
        "properties": {
          "team_id": {
            "type": "integer",
            "title": "Team Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "members": {
            "items": {
              "$ref": "#/components/schemas/RosterMember"
            },
            "type": "array",
            "title": "Members"
          }
        },
        "type": "object",
        "required": [
          "team_id",
          "name",
          "version",
          "members"
        ],
        "title": "TeamSnapshot"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}